VC-2376 adn VC-2372 update insights module to typescript and split class to separate file
