+++ package.json
@@ -25,6 +25,7 @@
     "npm": ">=5.8"
   },
   "devDependencies": {
+    "@types/lodash": "4.14.176",
     "color-blend": "git://github.com/wpbakery/color-blend.git#5dd059a6dbdc6754f53f4eca09800953534d8322",
     "combokeys": "3.0.1",
     "compression-webpack-plugin": "6.0.5",
@@ -42,8 +43,8 @@
     "react-textarea-autosize": "8.3.3",
     "reactcss": "1.2.3",
     "standard": "14.3.4",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.4.4",
+    "ts-loader": "9.2.6",
+    "typescript": "4.4.4",
     "vc-webpack-vendors": "3.0.5",
     "webfontloader": "1.6.28"
   },
